+++ .gulp/tests.js
@@ -4,19 +4,31 @@ import mobileFriendlyTest from 'mobile-friendly-test-npm'
 import htmlSpeed from 'html-speed'
 import cssTest from 'css-test-npm'
 
-let apiKey = ''
-let url = ''
+import dotenv from 'dotenv';
+dotenv.config();
+
+/**
+ * Mobile Friendly Test
+ */
+const mftApiKey = process.env.MFT_KEY || '';
+const mftUrl = process.env.PROXY_URL || '';
 const mobileTestRes = () =>
-  mobileFriendlyTest(url, apiKey)
+  mobileFriendlyTest(mftUrl, mftApiKey)
 
 
-apiKey = ''
-url = ''
+/**
+ * HTML Speed Test
+ */
+const hstApiKey = process.env.HST_KEY || '';
+const hstUrl = process.env.PROXY_URL || '';
 const htmlSpeedRes = () =>
-  htmlSpeed(url, apiKey)
+  htmlSpeed(hstUrl, hstApiKey)
 
 
-url = ''
-const cssTestRes = () => cssTest(url)
+/**
+ * CSS Test
+ */
+const cssUrl = process.env.PROXY_URL || '';
+const cssTestRes = () => cssTest(cssUrl)
 
 export { mobileTestRes, htmlSpeedRes, cssTestRes }